refactor(watch): extract useWatchSessionMeta (session stamp + last channel identity)
